backend/pyproject.toml
[tool.poetry]
name = "dakje"
version = "0.1.0"
description = ""
authors = ["tenzin <ten13zin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
botok = "^0.8.1"
symspellpy = "^6.7.0"
Django = "^3.2.4"
django-environ = "^0.4.5"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2.12.1"
pytest = "^6.2.2"
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
